add: make risk filter a component
